+++ cfa/dataops/__init__.py
@@ -5,6 +5,6 @@ format. Use the existing configuration files as a starting point. Validations
 will run on all configurations.
 """
 
-from .catalog import datacat, get_data
+from .catalog import datacat, get_data, list_datasets
 
-__all__ = [datacat, get_data]
+__all__ = [datacat, get_data, list_datasets]