+++ cfa/dataops/workflows/covid/funcs/import_data.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
-from cfa.scenarios.dataops.catalog import BlobEndpoint, get_data
-from cfa.scenarios.dataops.etl.utils import get_today_date
+from cfa.dataops.catalog import BlobEndpoint, get_data
+from cfa.dataops.etl.utils import get_today_date
 
 
 def import_vaccination_data(